Hark!  'Tis the first might commit of the general camp system into SVN!
Added the basic Camp::Master module, associated bin/* scripts for camp control (generation, removal, refresh, restart, etc.), the pgsql generic Postgres configuration template, and so on.

All of this is from lt3, which is a Rails-only area at this point, but it is where the multi-type camp system originated.  Interchange-focused stuff (presumably from v2) will need to be backed in, hopefully without breaking the Rails stuff.



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@158 0e478272-3c31-0410-889b-87d32de1d2c3