Multi-part commit related to FragranceNet work and the need for MySQL support.
1. Extend support for the MySQL as the main camp system database, to be determined by the DSN specified in camp-db-user.  This still needs to be tested.
2. Abstract out database configuration details a bit such that the "pg_" prefix becomes "db_", making for relative database agnosticism in the configuration keys.
3. Add support for auto-detection of the camp type's database type based on the existence db-specific directories ("pgsql" or "mysql") under the camp type skeleton.
4. Modify the postgresql.conf template configuration file to reflect new "db_" token prefixes in place of "pg_".

The Postgres stuff has been tested and verified on lt3 (and camps on lt3 have been updated to reflect the change).

To make your local camp system work with these changes, you'll need to edit the local-config of each camp type, as well as any database configuration files (such as the database.yml file for a Rails camp) to replace "pg_" with "db_".  Existing camps should be okay, just the camp type skeletons would require adjustment.

Yes, the changes we're witnessing are making the procedural design of Camp::Master all the more painful; it would be vastly preferable if it were using OOP at this point, particularly as we get into increasing annoyance in database logic.  This would be a reasonable thing to do as a refactoring a bit down the road, but not during this early stage.



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@170 0e478272-3c31-0410-889b-87d32de1d2c3