Fixed bugs in pgsql.sql camp schema, SSL cert creation
schema/pgsql.sql: the introduction of the vcs_types table and FK to it from camps
was not done correctly; needed a DROP TABLE command for vcs_types, and needed to
defined the vcs_type column in it as the primary key.

lib/Camp/Master.pm: the prepare_apache() sub was modified by mistake (by me) a few
revisions ago such that it looks to base_path() . '/etc/camp.key' for the private
SSL key rather than type_path() . '/etc/camp.key'; this was due to merging things
with the camp system on v2.  Resolved this problem by adding an _ssl_private_key()
sub that looks to the type path first, then to base path second if needed, to find
the SSL private key.  This fix should be compatible with v2, lt3, and fragrancenet
camps.



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@191 0e478272-3c31-0410-889b-87d32de1d2c3