Delete the deprecated mkcampdb utility
The mkcampdb utility has been deprecated; it is no longer relied upon by
mkcamp (as Camp::Master provides a prepare_database() sub to do all the work);
the refresh-camp "--db" option effectively performs the same thing and is more
accurate with the spirit of use.  Therefore, removed mkcampdb so it is no longer
maintained and the refresh-camp script can be the canonical one for updating
a camp's database.



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@192 0e478272-3c31-0410-889b-87d32de1d2c3