Add support for configuration token parsing in copy-paths.yml
The source and target paths of copy-paths.yml (in a given camp type directory)
can now -- optionally -- be parsed for configuration tokens.  See the new docs
provided in the diff for how to make use of this.

This allows the camp setup to do things like create symlinks to a particular
path within the camp (by using __CAMP_PATH__, for example, in the source path).



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@195 0e478272-3c31-0410-889b-87d32de1d2c3