Fix bug in database-type-sensitive code.
The _db_type_dispatcher function for handling certain logical steps
with different functions per database type was always using the
database type of the base camps database, rather than the database
type for the camp being created.  Fixed this and made a new version
that is just for the base camp database type, which means that the
camp system can truly work on MySQL and Postgres worked on the same
box, with either used at any point (base camp db or camp type db).



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@199 0e478272-3c31-0410-889b-87d32de1d2c3