Extend camp system version control options to support Git and increase configurability
* Added Git as an option for VCS type within the camp system.  This involves added a variety
of new functions to Camp::Master to support Git particulars, and modifying existing functions
to add Git logic branches or to abstract out logic that previously assumed an SVN central repo.
* Added documentation to Camp::Master to explain the Git, SVN, and SVK options, usage thereof,
and configuration possibilities.
* Added new configuration variables for controlling the repository location both for SVN and Git,
as well as for specifying the default VCS type for a given camp type.
* Modified mkcamp to eliminate SVN as the hardcoded default choice for VCS, such that it relies
on the camp type's default_vcs_type configuration variable instead.
* Commented out the $ENV{PATH} assignment that was added a few commits ago and causes problems
depending on your server setup/environment.

This makes the camp system considerably more flexible and is in keeping with the recent proclamation
that the SCM tool of choice for End Point is Git.



git-svn-id: svn+ssh://intranet.endpoint.com/var/svn/trunk/camps@225 0e478272-3c31-0410-889b-87d32de1d2c3