Do not set catalog during initialization with catalog_path($catalog).
When parse_all_files() is called during initialization, it calls
catalog_path($catalog), which, in turn, calls
set_catalog($catalog). This causes the catalog to be set to whichever
catalog is parsed last in parse_all_files().

This fix sets the catalog temporarily in parse_all_files() only for
the duration that it is needed, then it is unset. That will allow
the _default_catalog() function to fire, if desired, or the catalog
to be left unset, as designed.

Reviewed by Jon Jensen and Ethan Rowe.