Initial implementation/stubbing of Camp::Deployment and accompanying unit test.
Documented the bulk of the initial functionality we want in Camp::Deployment, a class for managing and manipulating camp system deployments (i.e. where the config and whatnot for a camps application lives).

A subset of the functionality is implemented; stuff for paths and initializing a deployment skeleton is there, though the config file stuff will need to wait for getting YAML serialization sorted in Camp::Moose::Object.  The unit test verifies the basics of what's there.