Initial implementation of port range class
The Camp::PortRange class will allow for management of port allocation per resource.

This initial prototype is pretty bare-bones and not really complete.  One problem is that the
range represents the metadata (the maximum range for all camps), but within an actual camp object
tree, we would want the range to be specific to that camp (with the range derived by default from the
camp's number and the metadata of the range).

Consequently, I think we have a big TODO here:
* revise this so Camp::PortRange is actually Camp::Moose::Attribute::PortRange, with the port details
  and allocation logic as properties of an attribute metadata object
* add a helper function to Camp::Moose called has_port_range() for easy building of attributes of this
  proposed type.
* have the proposed attribute type always be lazy by default and use a default sub that invokes some
  standard port allocation logic, so actual objects with these port range attributes get a simple list
  of port numbers for the attribute value.

That's significantly different from the way things are laid out in this commit, but the commit at least
gets us started with some design/logic/etc. we can use.