Allow remote-style git repo urls in local-config
Adds the logic to bring the code in line with the documentation by
allowing remote-hosted git repos.  We now detect a repo-path with a
protocol specifier or a user@host: protocol-less URL as belonging to a
remote git repo, and suppress adding the leading path information.

Bug reported by Adam Vollrath <adam@endpoint.com>