Reducing complexity in camp-info.
Simplified the mapping of options parsed by Getopt to database fields for camps for display.
Removed code related to option aliases that are not documented or are redundant and confusing.
There is no other field that makes sense to set except for comment, so removed unneccessary code abstracting that out.