Deprecate db_dbnames configuration; make ~/.pgpass match any database name
It only caused trouble to try to restrict which database any given Postgres
user could connect as. The actual security (if any) is set up on the server
side, and the client side might as well allow a Postgres user's password to
work with any database you care to try.